+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Upcoming
 * Fixed a crash and inconsistencies in the preview if both "Ignore folder structure" and "Selected in preview" are enabled. This also adversely causes the preview to be slower upon selecting items when "Selected in preview" is enabled. The drop in performance is only noticeable for hundreds of layers in an image.
 * Fixed export failing for some custom procedures modifying the number of layers, e.g. `gimp-image-merge-visible-layers`.
 * Built-in procedures and constraints are now listed alphabetically.
+* Renamed the "Rename layer" procedure to just "Rename" for brevity.
 * Updated documentation with examples on how to customize the export.